Add files via upload
In this file I have created a Dash application which uses plotly express in order to visualise various environmental data retrieved from the World Bank dataset for the years 2018-1990.

The application allows the user to select between:
a. Countries;
b. Sub-regions;
c. Continents

for:
a. year (between 1990 and 2018);
b. all years

There are 24 graphs in total.